requirements.txt
asgiref==3.4.1
autopep8==1.6.0
azure-core==1.23.1
azure-storage-blob==12.11.0
backports.zoneinfo==0.2.1
bson==0.5.8
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
dataclasses==0.6
Django==4.0.1
django-appconf==1.0.5
django-cors-headers==3.11.0
django-filter==21.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
djongo==1.3.6
et-xmlfile==1.1.0
Faker==13.3.2
idna==3.3
importlib-metadata==4.10.0
isodate==0.6.1
Markdown==3.3.6
msrest==0.6.21
numpy==1.22.3
oauthlib==3.2.0
openpyxl==3.0.9
pycodestyle==2.8.0
pycparser==2.21
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.2.4
toml==0.10.2
toposort==1.7
typing-extensions==4.1.1
urllib3==1.26.9
zipp==3.7.0
